requirements.txt
openai>=1.0.0
anthropic>=0.18.0
python-dateutil>=2.8.0
regex>=2023.0.0
pydantic>=2.0.0
